requirements.txt
asgiref==3.7.2
boto3==1.28.83
botocore==1.31.83
Django==4.2.7
django-cors-headers==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
install==1.3.5
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
whitenoise==6.6.0
